tools/telegram-mcp/requirements.txt
telethon>=1.40.0
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
